+++ test/unit/models/color-identity.test.ts
@@ -132,28 +132,4 @@ describe("ColorIdentity", () => {
       expect(ci.toString()).toEqual("g,b,r,w,u");
     });
   });
-
-  describe("toMarkdown", () => {
-    it("returns a markdown string", () => {
-      const ci = new ColorIdentity("w,r,b");
-
-      expect(ci.toMarkdown()).toEqual(":manaw::manab::manar:");
-    });
-  });
-
-  describe("toHTML", () => {
-    it("returns a document fragment", () => {
-      const ci = new ColorIdentity("w,r,b");
-
-      const fragment = ci.toHTML();
-      const imgs = fragment.querySelectorAll<HTMLImageElement>("img");
-
-      expect(fragment).toBeInstanceOf(DocumentFragment);
-
-      expect(imgs.length).toBe(3);
-      expect(imgs[0].src).toMatch("W.svg");
-      expect(imgs[1].src).toMatch("B.svg");
-      expect(imgs[2].src).toMatch("R.svg");
-    });
-  });
 });