+++ CHANGELOG.md
@@ -13,6 +13,7 @@ _Breaking Changes_
 - `CardGrouping`
   - Remove `matchesAll`
   - Remove `matchesAny`
+- Remove methods that return HTML Elements
 
 # 0.5.2